refactor: modularize home and playlists pages
