+++ frontend/src/services/movieApi.ts
@@ -1,0 +1,19 @@
+import type { Movie } from "../types";
+
+const API_URL = import.meta.env.VITE_API_URL ?? "http://localhost:3000";
+
+async function parseResponse<T>(response: Response): Promise<T> {
+  const data = await response.json();
+
+  if (!response.ok) {
+    throw new Error(data.message ?? "Erro inesperado");
+  }
+
+  return data;
+}
+
+export async function getMovies(): Promise<Movie[]> {
+  const response = await fetch(`${API_URL}/movies`);
+
+  return parseResponse<Movie[]>(response);
+}
